requirements.txt
alembic==1.11.1
appnope==0.1.3
asttokens==2.4.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.6
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.13.1
firebase-admin==6.5.0
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
fonttools==4.47.0
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
html5lib==1.1
httplib2==0.22.0
idna==3.4
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lxml==4.9.4
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
msgpack==1.0.8
nest-asyncio==1.5.9
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pyzmq==25.1.2
requests==2.31.0
retrying==1.3.4
rsa==4.9
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.19
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.25.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.3.6
